mvp 0.0.1 save

+++ backend/src/main/java/com/example/alabuga/service/ArtifactService.java
@@ -82,8 +82,8 @@ public class ArtifactService {
     public void deleteArtifact(Long id) {
         Artifact artifact = artifactRepository.findById(id)
                 .orElseThrow(() -> new ResourceNotFoundException("Артефакт", id));
-        
-        artifactRepository.delete(artifact);
+        artifact.setIsActive(false);
+        artifactRepository.save(artifact);
     }
     
     @Transactional
@@ -104,7 +104,6 @@ public class ArtifactService {
     }
     
     public List<UserArtifactDTO> getOtherUserArtifacts(Long userId) {
-        // Проверяем, что пользователь существует
         userRepository.findById(userId)
                 .orElseThrow(() -> new ResourceNotFoundException("Пользователь", userId));
         
@@ -119,9 +118,7 @@ public class ArtifactService {
         
         Artifact artifact = artifactRepository.findById(artifactId)
                 .orElseThrow(() -> new ResourceNotFoundException("Артефакт", artifactId));
-        
-        // Проверяем, не имеет ли уже пользователь этот артефакт
-        if (userArtifactRepository.findByUserIdAndArtifactId(userId, artifactId).isPresent()) {
+        if (userArtifactRepository.findByUserIdAndArtifactId(userId, artifactId) != null) {
             throw new BusinessLogicException("Пользователь уже имеет этот артефакт");
         }
         
@@ -137,16 +134,20 @@ public class ArtifactService {
     
     @Transactional
     public void removeArtifactFromUser(Long userId, Long artifactId) {
-        UserArtifact userArtifact = userArtifactRepository.findByUserIdAndArtifactId(userId, artifactId)
-                .orElseThrow(() -> new ResourceNotFoundException("Артефакт пользователя", artifactId));
+        UserArtifact userArtifact = userArtifactRepository.findByUserIdAndArtifactId(userId, artifactId);
+        if (userArtifact == null) {
+            throw new ResourceNotFoundException("Артефакт пользователя", artifactId);
+        }
         
         userArtifactRepository.delete(userArtifact);
     }
     
     @Transactional
     public UserArtifactDTO toggleArtifactEquip(Long userId, Long artifactId) {
-        UserArtifact userArtifact = userArtifactRepository.findByUserIdAndArtifactId(userId, artifactId)
-                .orElseThrow(() -> new ResourceNotFoundException("Артефакт пользователя", artifactId));
+        UserArtifact userArtifact = userArtifactRepository.findByUserIdAndArtifactId(userId, artifactId);
+        if (userArtifact == null) {
+            throw new ResourceNotFoundException("Артефакт пользователя", artifactId);
+        }
         
         userArtifact.setIsEquipped(!userArtifact.getIsEquipped());
         UserArtifact savedUserArtifact = userArtifactRepository.save(userArtifact);
